Create code
This code calculates the speed of the drone based on its position data at different points in time. Make sure to adapt the code to your specific use case and input data format. If you have access to real-time data from the drone, you may need to continuously update the position data and recalculate the speed.